eng-1211 Refactor exportUtils, factor out utility functions and markdown. This is pure refactoring with no change of functionality.
